+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ### documentdb v0.108-0 (Unreleased) ###
 * Top-level `let` variables and `$$NOW` supported by default.
-* Fix collation support on find and aggregation when variableSpec is not available.
+* Fix collation support on find and aggregation when variableSpec is not available *[Bugfix]*.
 * Support `dropRole` command *[Feature]*
 * Fix concurrent upsert behavior, update the documents in case of conflicts during insert *[Bugfix]* (#295).
+* Support collation with `$sortArray` aggregation operator *[Feature]*
 
 ### documentdb v0.107-0 (Unreleased) ###
 * Support sort by _id against the _id index using the enableIndexOrderbyPushdown flag *[Feature]*.